backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.1.8
dnspython==2.7.0
exceptiongroup==1.3.1
fastapi==0.128.4
h11==0.16.0
idna==3.11
motor==3.7.1
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
cloudinary==1.41.0
python-multipart==0.0.20
pypdf==6.6.2
google-generativeai==0.8.6
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
requests==2.32.4